Фикс make alias, добавлены алиасы на load post > post load, фикс уничтожения несостоявшейся PartActivity, pre в редакторе.

+++ src/com/cab404/acli/PartActivity.java
@@ -89,7 +89,8 @@ public class PartActivity extends AbstractActivity implements FragmentedList {
 	}
 
 	@Override protected void onDestroy() {
-		part.onRemove(root.getChildAt(0), root, this);
+		if (part != null && root != null)
+			part.onRemove(root.getChildAt(0), root, this);
 		super.onDestroy();
 	}
